GP2-3215: Shuffle custom StreamBlocks for better use elsewhere
