+++ great_international/blocks/investment_atlas.py
@@ -1,57 +1,13 @@
 """Streamfield blocks for Investment Atlas pages"""
 
-from django import forms
-
 from wagtail.core import blocks
-from wagtail.images.blocks import ImageChooserBlock
 from wagtail.snippets.blocks import SnippetChooserBlock
 
-from core.blocks import ButtonBlock
-from core.helpers import render_markdown
-from core.widgets import MarkdownTextarea
+from core.blocks import ButtonBlock, BaseAltTextImageBlock
+from great_international.blocks.great_international import MarkdownBlock
 
 
-DEFAULT_IMAGE_RENDITION_SPEC = "fill-960x540"
-
-
-class BaseAtlasImageBlock(blocks.StructBlock):
-    image = ImageChooserBlock(required=True)
-    image_alt = blocks.CharBlock(
-        max_length=255,
-        required=True,
-    )
-    caption = blocks.CharBlock(
-        max_length=255,
-        required=False,
-    )
-
-    def get_api_representation(self, value, context=None):
-        """Expand the default serialization to show the image rendition URL"""
-        if not context:
-            context = {}
-
-        retval_dict = {}
-
-        img = value.get("image")
-
-        if img:
-            _rendition_spec = context.get('rendition_spec', DEFAULT_IMAGE_RENDITION_SPEC)
-            retval_dict.update(
-                {
-                    'image': img.get_rendition(_rendition_spec).url,
-                    'image_alt': value.get('image_alt'),
-                    'caption': value.get('caption'),
-                }
-            )
-
-        return retval_dict
-
-
-class FeaturedImageBlock(BaseAtlasImageBlock):
-    pass
-
-
-class InlineOpportunityImageBlock(BaseAtlasImageBlock):
+class InlineOpportunityImageBlock(BaseAltTextImageBlock):
     custom_css_class = blocks.CharBlock(
         max_length=255,
         required=False,
@@ -63,24 +19,6 @@ class InlineOpportunityImageBlock(BaseAtlasImageBlock):
         if retval_dict:
             retval_dict.update({'custom_css_class': value.get('custom_css_class')})
         return retval_dict
-
-
-class MarkdownBlock(blocks.FieldBlock):
-    def __init__(self, required=True, help_text=None, **kwargs):
-        self.field = forms.CharField(
-            required=required,
-            help_text=help_text,
-            widget=MarkdownTextarea(),
-        )
-        super().__init__(**kwargs)
-
-    def get_api_representation(self, value, context=None):
-        """Expand the default serialization to turn the Markdown into HTML"""
-        raw_markdown = super().get_api_representation(value, context)
-        if raw_markdown:
-            return render_markdown(raw_markdown)
-        else:
-            return ''
 
 
 # This is separate because it's used in both PageSectionBlock, below, and
@@ -132,7 +70,7 @@ class ContactDetailsBlock(blocks.StructBlock):
     )
 
 
-class AtlasLandingPagePanelImageBlock(BaseAtlasImageBlock):
+class AtlasLandingPagePanelImageBlock(BaseAltTextImageBlock):
     pass
